requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-aws>=0.2.0
langchain-community>=0.3.0
langgraph>=0.3.0
langsmith>=0.2.0
tavily-python>=0.5.0
pypdf>=4.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-pptx>=1.0.0
beautifulsoup4>=4.12.0
